+++ src/main/java/net/jptrzy/inventory/backpack/config/ModConfig.java
@@ -1,0 +1,5 @@
+package net.jptrzy.inventory.backpack.config;
+
+public class ModConfig {
+    public static boolean trinkets = true;
+}